// src/models/CovidFinder.ts
export interface CovidFinder {
  loc_phone: string;
  loc_name: string;
  loc_admin_street1: string;
  loc_admin_city: string;
  loc_admin_state: string;
  loc_admin_zip: string;
  web_address: string;
  pre_screen: string;
  insurance_accepted: boolean;
  walkins_accepted: boolean;
  med_name: string;
  in_stock: boolean;
  quantity_last_updated: string;
}
// export interface CovidFinder {
//     provider_location_guid: string;
//     loc_store_no: string;
//     loc_phone: string;
//     loc_name: string;
//     loc_admin_street1: string;
//     loc_admin_city: string;
//     loc_admin_state: string;
//     loc_admin_zip: string;
//     sunday_hours: string;
//     monday_hours: string;
//     tuesday_hours: string;
//     wednesday_hours: string;
//     thursday_hours: string;
//     friday_hours: string;
//     saturday_hours: string;
//     web_address: string;
//     pre_screen: string;
//     insurance_accepted: boolean;
//     walkins_accepted: boolean;
//     provider_notes: string;
//     ndc: string;
//     med_name: string;
//     in_stock: boolean;
//     supply_level: string;
//     quantity_last_updated: string;
//     latitude: string;
//     longitude: string;
//     category: string;
//     offers_free_masks: boolean;
//     min_age_months: string;
//     min_age_years: string;
//   }